Fix first run initialization of zwallet with encrypted wallet.dat file
When running the client with a `wallet.dat` file that has never
initialized it's zwallet or dzPIV, there is an incorrect code pathway
that leads to the dzPIV seed hash never being written to the wallet
file. Instead, we copy the seed generation code into `crypter.cpp` which
will allow for proper dzPIV seed generation upon first unlock.