[QT] Fix a display bug about zPIV mints
When opening the zPIV coin selection in privacy tab, if the wallet is
locked the message for all mints is that the seed is wrong. This commit
fixes it by replacing the message and stating that the locked wallet
prevents precomputation and spending.

Github-Pull: #852
Rebased-From: 94bea35e5569c323a667d8204bc2d23b15d4dc29