Mining Fixes (#37)
* Give error if pow phase is finished

* Force Sync to allow for staking without mnsync

* Fix error

* Update rpcmisc.cpp

* Update rpcmisc.cpp

* more ignore

* Update fundamentalnode-sync.cpp

* Add gitian build python script

* Update Univalue

* add missing gencpp

* Add missing includes

* Fix Gitian build

Update Univalue


add missing gencpp


Add missing includes

* Revert "Update fundamentalnode-sync.cpp"

This reverts commit 323b13fc599b03982b0ca7c3f32342537a6b992f

* Fix gitian build and minor updates

* remove forcesync

* [Model] Invalid iterator position crash fix.

* background loading crash fixed.

* random: fix crash on some 64bit platforms

rbx needs to be stashed in a 64bit register on 64bit platforms. With this crash
in particular, it was holding a stack canary which was not properly restored
after the cpuid.

Split out the x86+PIC case so that x86_64 doesn't have to worry about it.

* random: Add fallback if getrandom syscall not available

If the code was compiled with newer (>=3.17) kernel headers but executed
on a system without the system call, every use of random would crash the
program. Add a fallback for that case.

* Don't assert if we were beaten to the block

A timing window exists where a wallet could be creating a new block from within the miner thread when a new block is received to the wallet.  This window will create a situation where TestBlockValidity() fails because the chain tip has changed between the time it created the new block and the time it tested the validity of the block.

This situation would result in the wallet being asserted; however this is a little overkill.  rather than asserting if the tip has changed, it is better to throw the block away.

This problem was revealed during a testnet test of an altcoin, and very prevalent when multiple wallet existed with the exact same number of staking coins received in the same transaction; or when multiple wallets were staking the same coins via import private key.  The problem happens significantly less in more normal circumstances, but was still observed in a testing environment with fast blocks.

It is likely that this scenario has been encountered but never determined to be root cause, as a crashed wallet could be restarted, re-indexed and never investigated further.

* [Wallet] Add some LOCK to avoid crash

Github-Pull: #625
Rebased-From: ed23d0d62054156e0c2cb632fecbd65588c50e8c

* testnet mining fixes