[Build] Disable apt-cacher for Windows WSL gitian setup
gitian's use of apt-cacher server on the host doesn't work when using
Windows 10's WSL subsystem. This adds a check to see if the ubuntu OS
version is from "Microsoft" (which indicates it is a WSL environment),
then uses the boolean result to disable the use of apt-cacher if true.

Github-Pull: #1536
Rebased-From: 45ce02e1e0d93b90eadfd954599829c8c7e2b111